refactor packing components
