+++ src/pack/index.ts
@@ -1,0 +1,2 @@
+export { Packer } from './packer';
+export type { PackResultMetadata, PackWriteOptions, PackOptions, PackResult } from './packer';
